move to separate hook and cleanup
